pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "pip>=21.0"]
build-backend = "setuptools.build_meta"

[project]
name = "soundslike"
version = "0.1.0"
description = "A library for sonifying probability distributions and statistical concepts"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "numpy>=1.20.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "sound-machine>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "jupyter>=1.0.0",
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=soundslike"
testpaths = ["tests"] 